+++ documentation/docs/release-notes.md
@@ -6,7 +6,8 @@ title: Release notes&#58;
 **v5.0.1**:
 
 - Hazelcast-based implementation for SAMLMessageStore
-- Added an option to tolerate missing SAML response `Destination` attribute.
+- Added an option to tolerate missing SAML response `Destination` attribute
+- SAML support: don't add the friendly name if it is the same as the (mapped) name (avoid duplication of values)
 
 **v5.0.0** (see: [what's new in pac4j v5?](/blog/what_s_new_in_pac4j_v5.html)):